e2e: define Docker image (cip-e2e-env) that can run e2e tests
This is in preparation for creating a Prow job that can run e2e tests
for the promoter.

The way we generate the cip-e2e-env image is a bit hacky, but it's
better than writing a raw Dockerfile because the images are
reproducible. The image has:

    Bazel
    Docker
    GCloud SDK
    Golang
    make

The first 3 are taken care of by
https://github.com/GoogleCloudPlatform/container-definitions/tree/80de5376874035319403400fa253b2785c31f3e2/ubuntu1604_bazel_docker_gcloud.

The golang compiler installation is taken from
https://github.com/GoogleCloudPlatform/layer-definitions/tree/a8dfda9de289c68d217a15904f8a6500dcce7c88.

Lastly the "make" binary is taken from a Debian Jessie snapshot.